+++ src/main/scala/csvUtils.scala
@@ -19,7 +19,7 @@ case object csvUtils {
 }
 
 case class RowOps(row: csvUtils.Row) extends AnyVal {
-  import csvUtils._, RNACentral5._
+  import csvUtils._, rnacentral.RNACentral5._
 
   def toMap: Map[Field, String] = Id2Taxa.keys.types.asList.zip(row).toMap
