+++ src/main/scala/rnacentral/compats.scala
@@ -1,9 +1,9 @@
-package era7bio.db
+package era7bio.db.rnacentral
 
 import ohnosequences.statika._, aws._
 import ohnosequences.awstools._, regions.Region._, ec2._, InstanceType._, autoscaling._, s3._
 
-case object rnaCentralCompats {
+case object compats {
 
   class DefaultCompatible[B <: AnyBundle](bundle: B, javaHeap: Int = 10 /*G*/) extends Compatible(
     amznAMIEnv(AmazonLinuxAMI(Ireland, HVM, InstanceStore), javaHeap),
@@ -11,5 +11,5 @@ case object rnaCentralCompats {
     generated.metadata.db.rnacentral
   )
 
-  case object MirrorRNAcentral5 extends DefaultCompatible(era7bio.db.MirrorRNAcentral5)
+  case object MirrorRNAcentral5 extends DefaultCompatible(era7bio.db.rnacentral.MirrorRNAcentral5)
 }
